+++ UHDD/src/application/LoginResult.java
@@ -1,0 +1,7 @@
+package application;
+
+public enum LoginResult {
+		SUCCESSFUL,
+	    WRONG_CODE,
+	    CANCELLED
+	}
