+++ UHDD/src/application/viewControllers/pwdResetController.java
@@ -1,6 +1,7 @@
 package application.viewControllers;
 
 import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -8,6 +9,7 @@ import com.mysql.cj.protocol.Resultset;
 
 import application.CredentialManager;
 import application.DBConnector;
+import application.LoginResult;
 import application.PasswordHash;
 import application.PasswordHasher;
 import javafx.animation.PauseTransition;
@@ -15,6 +17,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
+import javafx.scene.paint.Color;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
 import javafx.util.Duration;
@@ -36,7 +39,7 @@ public class pwdResetController {
   
 
     
-    public void resetPwd() throws SQLException, ClassNotFoundException {
+    public void resetPwd() throws SQLException, ClassNotFoundException, NoSuchAlgorithmException {
         String oldPwd = txtOldPwd.getText();
         String newPwd = txtNewPwd.getText();
         String verifyPwd = txtVerifyPwd.getText();
@@ -54,17 +57,30 @@ public class pwdResetController {
         } else if (newPwd.length() < 8 || newPwd.length() > 64) {
             txtStatus.setText("New password must be between 8 and 64 characters");
         } else if (credentialManager.verifyPassword(username, oldPwd)){
-            credentialManager.changePasswordInDB(username, newPwd);
-            txtStatus.setText("Password Updated");
-            txtStatus.setStroke(javafx.scene.paint.Color.GREEN);
-
-            PauseTransition delay = new PauseTransition(Duration.seconds(2)); // Creates a 2 seconds pause
-            delay.setOnFinished( event -> {
-                // Closes the window after the pause
-                Stage stage = (Stage) txtStatus.getScene().getWindow(); 
-                stage.close();
-            });
-            delay.play();
+            LoginResult result = credentialManager.verifyMFA(username);
+            if (result == LoginResult.SUCCESSFUL) {
+                // Handle successful login
+                credentialManager.changePasswordInDB(username, newPwd);
+                txtStatus.setText("Password Updated");
+                txtStatus.setStroke(javafx.scene.paint.Color.GREEN);
+                PauseTransition delay = new PauseTransition(Duration.seconds(2)); // Creates a 2 seconds pause
+                delay.setOnFinished( event -> {
+                    // Closes the window after the pause
+                    Stage stage = (Stage) txtStatus.getScene().getWindow(); 
+                    stage.close();
+                });
+                delay.play();               
+            } else if (result == LoginResult.WRONG_CODE) {
+                txtStatus.setText("Wrong code input. MFA verification cancelled");
+                txtStatus.setFill(Color.RED);        
+            } else if (result == LoginResult.CANCELLED) {
+                // Handle login cancelled scenario
+                txtStatus.setText("MFA verification cancelled");
+                txtStatus.setFill(Color.RED);  
+            } else {
+                // Handle any other result if necessary
+               
+            }     
         }
         else {
             txtStatus.setText("Old password is incorrect");
@@ -72,7 +88,7 @@ public class pwdResetController {
     }
 
     @FXML
-    protected void handlePwdResetAction(ActionEvent event) throws IOException, ClassNotFoundException, SQLException{
+    protected void handlePwdResetAction(ActionEvent event) throws IOException, ClassNotFoundException, SQLException, NoSuchAlgorithmException{
         resetPwd();
     }
 }
